docs: add guidelines and backend READMEs; consolidate sentry init
