+++ src/pages/ApiDocsHub.tsx
@@ -1,6 +1,6 @@
 
 import React from "react";
-import { Link } from "react-router-dom";
+import Link from "next/link";
 import { ArrowRight, BookOpen, Code, Webhook, AlertCircle, Zap } from "lucide-react";
 import ApiDocsLayout from "@/components/developers/ApiDocsLayout";
 
@@ -53,9 +53,9 @@ export function ApiDocsHub() {
 
         <div className="grid gap-6 md:grid-cols-2">
           {sections.map((section) => (
-            <Link 
-              key={section.path} 
-              to={section.path}
+            <Link
+              key={section.path}
+              href={section.path}
               className="block p-6 rounded-lg bg-zinc-900 border border-zinc-800 hover:border-zinc-700 transition-all"
             >
               <div className={`w-12 h-12 rounded-full bg-gradient-to-r ${section.color} flex items-center justify-center mb-4`}>
@@ -77,7 +77,7 @@ export function ApiDocsHub() {
             To use the Zion AI Marketplace API, you'll need an API key. Visit your developer portal to create one.
           </p>
           <Link
-            to="/developers/portal"
+            href="/developers/portal"
             className="inline-flex items-center px-4 py-2 rounded-md bg-zion-purple text-white hover:bg-zion-purple/90 transition-colors"
           >
             Go to Developer Portal
