+++ src/pages/Unauthorized.tsx
@@ -1,5 +1,5 @@
 
-import { Link } from "react-router-dom";
+import Link from "next/link";
 import { Button } from "@/components/ui/button";
 import { Header } from "@/components/Header";
 import { Footer } from "@/components/Footer";
@@ -22,13 +22,13 @@ export default function Unauthorized() {
           </p>
           <div className="flex flex-col gap-3">
             <Link
-              to="/"
+              href="/"
               className="w-full bg-gradient-to-r from-zion-purple to-zion-purple-dark hover:from-zion-purple-light hover:to-zion-purple text-white px-4 py-2 rounded inline-flex items-center justify-center"
             >
               Return to Home
             </Link>
             <Link
-              to="/login"
+              href="/login"
               className="w-full border border-zion-blue-light text-zion-slate-light hover:bg-zion-blue hover:text-white px-4 py-2 rounded inline-flex items-center justify-center"
             >
               Login with Different Account
